[diffusion] Move qknorm_rope JIT kernel under diffusion layout
