+++ python/sglang/jit_kernel/benchmark/diffusion/bench_qknorm_rope.py
@@ -139,7 +139,7 @@ def split_qknorm_rope(inputs: dict[str, torch.Tensor | bool]) -> None:
 
 
 def fused_qknorm_rope(inputs: dict[str, torch.Tensor | bool]) -> None:
-    from sglang.jit_kernel.qknorm_rope import fused_inplace_qknorm_rope
+    from sglang.jit_kernel.diffusion.qknorm_rope import fused_inplace_qknorm_rope
 
     fused_inplace_qknorm_rope(
         inputs["q"],
