+++ python/sglang/jit_kernel/tests/diffusion/test_qknorm_rope.py
@@ -69,7 +69,7 @@ def fused_qknorm_rope(
     positions: torch.Tensor,
     is_neox: bool,
 ) -> None:
-    from sglang.jit_kernel.qknorm_rope import fused_inplace_qknorm_rope
+    from sglang.jit_kernel.diffusion.qknorm_rope import fused_inplace_qknorm_rope
 
     fused_inplace_qknorm_rope(
         q,
